% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_sen.R
\name{volcano_sen}
\alias{volcano_sen}
\title{Create a volcano plot}
\usage{
volcano_sen(
  table,
  title = "Cond_Up vs Cond_Down",
  significanceMeasure = c("fdr", "PValue"),
  thresSig = 0.05,
  thresLogFC = 0.58,
  colorUp = "darkgreen",
  colorDown = "steelblue",
  colorNeutral = "lightgrey",
  xlim = c(-10, 10),
  ylim = c(0, 15)
)
}
\arguments{
\item{table}{Input table with logFC and (fdr or PValue) columns}

\item{title}{Title of the plot}

\item{significanceMeasure}{the significance Measure to be used for statistical significance. Must match the corresponding column' name}

\item{thresSig}{The statistical Measure significance value}

\item{thresLogFC}{The Log Fold change significance threshold}

\item{colorUp}{The upregulated genes colors}

\item{colorDown}{The downregulated genes colors}

\item{colorNeutral}{The color of non-statistically significant or non-logFC significant genes}
}
\value{
ggplot object
}
\description{
Creates a volcano plot based on a gene list with logFC and (fdr or PValue) columns
Works best with edgeR results.
}
\details{
date 24/10/2022
}
